+++ super_admin_1/notification/config.py
@@ -1,0 +1,3 @@
+url_mapping = {
+    "unban": "/api/v1/store/suspension-lifted"
+}
