+++ super_admin_1/notification/notification_helper.py
@@ -2,7 +2,8 @@
 import requests
 from typing import Dict
 from super_admin_1.models.alternative import Database
-from super_admin_1.products.product_action_logger import logger
+from super_admin_1.logs.product_action_logger import logger
+from super_admin_1.notification.config import url_mapping
 
 
 class CustomError(Exception):
@@ -55,24 +56,25 @@ def notify(vendor_id: str, action: str, **kwargs: str) -> dict:
             - data (dict): details of the successful mail sent
             - error (bool): signifies error during execution
     """
-    email_request_url = "https://team-titan.mrprotocoll.me/api/v1/mail/test-email"
+    email_request_base_url = "https://team-titan.mrprotocoll.me/api/v1/mail/test-email"
     try:
         # query the database to get the recipient email
-        email = get_field_value(field="email", table="public.user",
+        email, name = get_field_value(field="email, name", table="public.user",
                                 filter="id", value=vendor_id)
         data: Dict = {
             "recipient": email,
-            "action": action
+            "name": name
+            # "action": action
         }
         # update the data dictionary with the available keys
         if kwargs.get("product_id", None):
             product_name = get_field_value(field="name", table="product",
-                                        filter="id", value=kwargs.get("product_id"))
+                                           filter="id", value=kwargs.get("product_id"))
             data["product_name"] = product_name
         else:
             shop_name = get_field_value(field="name", table="shop",
                                         filter="id", value=kwargs.get("shop_id"))
-            data["shop_name"] = shop_name
+            data["store_name"] = shop_name
         if kwargs.get("reason", None):
             data["reason"] = kwargs.get("reason")
         
@@ -80,7 +82,8 @@ def notify(vendor_id: str, action: str, **kwargs: str) -> dict:
         logger.error(f"{type(error).__name__}: {error}")
 
     try:
-        response = requests.post(email_request_url, json=data)
+        endpoint = url_mapping.get(action)
+        response = requests.post(f"{email_request_base_url}/{endpoint}", json=data)
         if response.status_code != 200:
             return {
                 "success": False,
@@ -107,15 +110,16 @@ def notify(vendor_id: str, action: str, **kwargs: str) -> dict:
 
 def notify_test(name: str, email: str) -> dict:
     
-    email_request_url = "https://team-titan.mrprotocoll.me/api/v1/assessment/badge"
+    email_request_url = "https://team-titan.mrprotocoll.me/api/v1/store/suspension-lifted"
     try:
         
         data: Dict = {
             "name": name,
             "recipient": email,
-            "skill": "Content Writer",
-            "badge_name": "Content Writing",
-            "user_profile_link": "https://example.com"
+            "store_name": "Okay store"
+            # "skill": "Content Writer",
+            # "badge_name": "Content Writing",
+            # "user_profile_link": "https://example.com"
         }
        
     except Exception as error:
