+++ super_admin_1/shop/routes.py
@@ -584,75 +584,6 @@ def get_temporarily_deleted_vendors():
         return jsonify({"status": "Error", "message": str(e)})
 
 
-logs = Blueprint("logs", __name__, url_prefix="/api/logs")
-
-
-@logs.route("/shops", defaults={"shop_id": None})
-@logs.route("/shops/<shop_id>")
-@super_admin_required
-def get_all_shop_logs(shop_id):
-    """Get all shop logs"""
-    if not shop_id:
-        return (
-            jsonify(
-                {
-                    "message": "success",
-                    "logs": [
-                        log.format() if log else [] for log in ShopsLogs.query.all()
-                    ],
-                }
-            ),
-            200,
-        )
-
-    return (
-        jsonify(
-            {
-                "message": "success",
-                "logs": [
-                    log.format() if log else []
-                    for log in ShopsLogs.query.filter_by(shop_id=shop_id).all()
-                ],
-            }
-        ),
-        200,
-    )
-
-
-@logs.route("/shops/download", defaults={"shop_id": None})
-@logs.route("/shops/<shop_id>/download")
-@super_admin_required
-def download_shop_logs(shop_id):
-    """Download all shop logs"""
-    logs = []
-    if not shop_id:
-        logs = [log.format() if log else [] for log in ShopsLogs.query.all()]
-    else:
-        logs = [
-            log.format() if log else []
-            for log in ShopsLogs.query.filter_by(shop_id=shop_id).all()
-        ]
-    # Create a temporary file to store the strings
-    temp_file_path = f"{os.path.abspath('.')}/temp_file.txt"
-    with open(temp_file_path, "w") as temp_file:
-        temp_file.write("\n".join(logs))
-
-    response = send_file(
-        temp_file_path, as_attachment=True, download_name="shoplogs.txt"
-    )
-    os.remove(temp_file_path)
-
-    return response
-
-
-@logs.route("/shop/actions", methods=["GET"])
-@super_admin_required
-def shop_actions():
-    data = ShopsLogs.query.all()
-    return jsonify([action.format_json() for action in data]), 200
-  
-  
-
 @shop.route("/sanctioned", methods=["GET"])
 # @super_admin_required
 def sanctioned_shop():
